+++ lib/kitchen/directions/move_solutions_to_answer_key/v1.rb
@@ -11,6 +11,8 @@ module Kitchen::Directions::MoveSolutionsToAnswerKey
           Strategies::UPhysics
         when :american_government
           Strategies::AmericanGovernment
+        when :precalculus
+          Strategies::Precalculus
         else
           raise 'No such strategy'
         end